+++ src/main/java/org/solstice/rollingStones/mixin/LivingEntityMixin.java
@@ -11,11 +11,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(LivingEntity.class)
 public class LivingEntityMixin {
 
-	@Inject(method = "createLivingAttributes", at = @At("RETURN"), cancellable = true)
+	@Inject(method = "createLivingAttributes", at = @At("RETURN"))
 	private static void appendLivingAttributes(CallbackInfoReturnable<DefaultAttributeContainer.Builder> cir) {
-		DefaultAttributeContainer.Builder builder = cir.getReturnValue();
-		builder.add(RollingAttributes.MAX_PULL_TIME_MULTIPLIER);
-		cir.setReturnValue(builder);
+		cir.getReturnValue().add(RollingAttributes.MAX_PULL_TIME_MULTIPLIER);
 	}
 
 }