+++ src/main/java/org/solstice/rollingStones/mixin/test/MobEntityMixin.java
@@ -9,10 +9,8 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.world.World;
-import org.solstice.rollingStones.registry.RollingTags;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 @Mixin(MobEntity.class)
 public abstract class MobEntityMixin extends LivingEntity {
@@ -21,27 +19,40 @@ public abstract class MobEntityMixin extends LivingEntity {
 		super(entityType, world);
 	}
 
-//	@WrapOperation(
-//		method = "tryAttack",
-//		at = @At(
-//			value = "INVOKE",
-//			target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"
-//		)
-//	)
-//	private boolean ignoreAttackWithTrainingWeapon(
-//		Entity instance,
-//		DamageSource source,
-//		float amount,
-//		Operation<Boolean> original
-//	) {
-//		if (this.getMainHandStack().isIn(RollingTags.TRAINING_WEAPON)) return true;
-//		return original.call(instance, source, amount);
-//	}
+	@WrapOperation(
+		method = "tryAttack",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"
+		)
+	)
+	private boolean ignoreDamageWithTrainingWeapon(
+		Entity instance,
+		DamageSource source,
+		float amount,
+		Operation<Boolean> original,
+		@Local(name = "f") float dealtDamage
+	) {
+		if (dealtDamage == 0F) return true;
+		return original.call(instance, source, amount);
+	}
 
-//	@ModifyVariable(method = "tryAttack", at = @At("STORE"), name = "g")
-//	private float increaseKnockbackWithTrainingWeapon(float strength) {
-//		if (this.getMainHandStack().isIn(RollingTags.TRAINING_WEAPON)) return ++strength;
-//		return strength;
-//	}
+	@WrapOperation(
+		method = "tryAttack",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/entity/mob/MobEntity;getKnockbackAgainst(Lnet/minecraft/entity/Entity;Lnet/minecraft/entity/damage/DamageSource;)F"
+		)
+	)
+	private float increaseKnockbackWithTrainingWeapon(
+		MobEntity instance,
+		Entity entity,
+		DamageSource damageSource,
+		Operation<Float> original,
+		@Local(name = "f") float dealtDamage
+	) {
+		if (dealtDamage == 0F) return 1;
+		return original.call(instance, entity, damageSource);
+	}
 
 }